a bit of clean up on experiments
