+++ Cargo.toml
@@ -17,7 +17,6 @@ grpc = "0.6.1"
 protobuf = "2"
 serde = "1.0.91"
 serde_json = "1.0.39"
-tower = {git = "https://github.com/tower-rs/tower", branch = "v0.3.x"}
 
 [build-dependencies]
 protoc-rust-grpc = "0.6.1"
